Selenium hw1 new structure, fixed tests
