+++ pom.xml
@@ -10,10 +10,10 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>16</maven.compiler.source>
-        <maven.compiler.target>16</maven.compiler.target>
+        <maven.compiler.source>15</maven.compiler.source>
+        <maven.compiler.target>15</maven.compiler.target>
         <testng.version>7.4.0</testng.version>
-        <assertj-core.version>3.17.0</assertj-core.version>
+        <assertj-core.version>3.20.2</assertj-core.version>
         <calculator.version>1.0</calculator.version>
         <maven-surefire-plugin.suiteXmlFile>testng</maven-surefire-plugin.suiteXmlFile>
         <maven-surefire-plugin.version>3.0.0-M5</maven-surefire-plugin.version>
@@ -26,7 +26,7 @@
             <artifactId>calculator</artifactId>
             <version>${calculator.version}</version>
             <scope>system</scope>
-            <systemPath>${basedir}/src/Selenium/test/resources/calculator-1.0.jar</systemPath>
+            <systemPath>${project.basedir}/src/test/resources/calculator-1.0.jar</systemPath>
         </dependency>
         <dependency>
             <groupId>org.testng</groupId>
@@ -50,6 +50,7 @@
                 <version>${maven-surefire-plugin.version}</version>
                 <configuration>
                     <suiteXmlFiles>
+                        ${maven-surefire-plugin.suiteXmlFile}/seleniumHW1addSubtractTests.xml
                     </suiteXmlFiles>
                 </configuration>
             </plugin>
@@ -58,59 +59,30 @@
 
     <profiles>
         <profile>
-            <id>addSubtract</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>${maven-surefire-plugin.version}</version>
-                        <configuration>
-                            <suiteXmlFiles>
-                                <suiteXmlFile>${maven-surefire-plugin.suiteXmlFile}/addSubtractTests.xml</suiteXmlFile>
-                            </suiteXmlFiles>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
+            <id>Selenium_hw1_addSubtract</id>
+            <properties>
+                <maven-surefire-plugin.suiteXmlFile>
+                    ${maven-surefire-plugin.suiteXmlFile}/seleniumHW1addSubtractTests.xml
+                </maven-surefire-plugin.suiteXmlFile>
+            </properties>
         </profile>
+
         <profile>
-            <id>multiplyDivision</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>${maven-surefire-plugin.version}</version>
-                        <configuration>
-                            <suiteXmlFiles>
-                                <suiteXmlFile>${maven-surefire-plugin.suiteXmlFile}/multDivisionTests.xml</suiteXmlFile>
-                            </suiteXmlFiles>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
+            <id>Selenium_hw1_multiplyDivisiont</id>
+            <properties>
+                <maven-surefire-plugin.suiteXmlFile>
+                    ${maven-surefire-plugin.suiteXmlFile}/seleniumHW1multDivisionTests.xml
+                </maven-surefire-plugin.suiteXmlFile>
+            </properties>
         </profile>
+
         <profile>
-            <id>all</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>${maven-surefire-plugin.version}</version>
-                        <configuration>
-                            <suiteXmlFiles>
-                                <suiteXmlFile>${maven-surefire-plugin.suiteXmlFile}/allTests.xml</suiteXmlFile>
-                            </suiteXmlFiles>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
+            <id>Selenium_hw1_all</id>
+            <properties>
+                <maven-surefire-plugin.suiteXmlFile>
+                    ${maven-surefire-plugin.suiteXmlFile}/seleniumHW1allTests.xml
+                </maven-surefire-plugin.suiteXmlFile>
+            </properties>
         </profile>
     </profiles>
-
 </project>
