--- src/Selenium/test/resources/testng/addSubtractTests.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<suite name="addSubtractSuite">
-    <test name="addSubtractTest">
-        <groups>
-            <run>
-                <include name="addSubtract"/>
-            </run>
-        </groups>
-        <classes>
-            <class name="hw1.AddTests"/>
-            <class name="hw1.SubtractTests"/>
-        </classes>
-    </test>
-</suite>
