--- src/Selenium/test/resources/testng/multDivisionTests.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<suite name="multiplyDivisionSuite">
-    <test name="multiplyDivisionTest">
-        <groups>
-            <run>
-                <include name="multiplyDivision"/>
-            </run>
-        </groups>
-        <classes>
-            <class name="hw1.MultiplyTests"/>
-            <class name="hw1.DivisionTests"/>
-        </classes>
-    </test>
-</suite>
