+++ src/test/java/Selenium_hw1/DivisionTests.java
@@ -1,4 +1,4 @@
-package hw1;
+package Selenium_hw1;
 import com.epam.tat.module4.Calculator;
 import org.assertj.core.api.Assertions;
 import org.testng.annotations.AfterClass;
@@ -7,6 +7,7 @@ import org.testng.annotations.Test;
 
 public class DivisionTests {
     private Calculator calculator;
+    private double EPS = 0.00001;
 
     @BeforeClass
     void initCalculator() {
@@ -18,7 +19,7 @@ public class DivisionTests {
         calculator = null;
     }
 
-    @Test(dataProvider = "longData", dataProviderClass = DataProviderCalc.class, groups = "multiplyDivision")
+    @Test(dataProvider = "longData", dataProviderClass = DataProviderCalc.class)
     public void divLongTest(long a, long b) {
         if (b == 0L) {
             Assertions.assertThatThrownBy(() -> calculator.div(a, b))
@@ -30,13 +31,14 @@ public class DivisionTests {
         }
     }
 
-    @Test(dataProvider = "doubleData", dataProviderClass = DataProviderCalc.class, groups = "multiplyDivision")
+    @Test(dataProvider = "doubleData", dataProviderClass = DataProviderCalc.class)
     public void divDoubleTest(double a, double b) {
         if (b == 0.0) {
             Assertions.assertThatThrownBy(() -> calculator.div(a, b)).isInstanceOf(NumberFormatException.class);
         } else {
             Assertions.assertThat(calculator.div(a, b))
-                    .as("Error in division double").isEqualTo(a / b);
+                    .as("Error in division double")
+                    .isCloseTo(a / b, Assertions.offset(EPS));
         }
     }
 }
